Move ImageService under Storage namespace and rename some spec files.
