+++ lib/azure/armrest/model/base_model.rb
@@ -182,7 +182,6 @@ module Azure
     class AvailabilitySet < BaseModel; end
     class DeploymentTemplate < BaseModel; end
     class Event < BaseModel; end
-    class Image < BaseModel; end
     class ImageVersion < BaseModel; end
     class Offer < BaseModel; end
     class Publisher < BaseModel; end
@@ -253,6 +252,7 @@ module Azure
 
     module Storage
       class Disk < BaseModel; end
+      class Image < BaseModel; end
       class Snapshot < BaseModel; end
     end
   end
