+++ spec/storage_disk_service_spec.rb
