+++ spec/storage_snapshot_service_spec.rb
